Cleanup and refactoring component
